Ready for Production

+++ src/krsuppliers/models/Purchase.java
@@ -4,7 +4,7 @@ import java.sql.Date;
 
 public class Purchase extends Transaction{
 
-    public Purchase(int _id, Date date, int bill, int particular_id, String particular, int qty, float rate, float discount, float amount){
+    public Purchase(int _id, Date date, int bill, int particular_id, String particular, float qty, float rate, float discount, float amount){
         super(_id, date, bill, particular_id, particular, qty, rate, discount, amount);
         super.setCategory(Category.PURCHASE);
     }
